+++ include/rocksdb/utilities/options_type.h
@@ -45,6 +45,7 @@ enum class OptionType {
   kConfigurable,
   kCustomizable,
   kEncodedString,
+  kCacheTierType,
   kUnknown,
 };