+++ options/options.cc
@@ -19,6 +19,7 @@
 #include "rocksdb/cache.h"
 #include "rocksdb/compaction_filter.h"
 #include "rocksdb/comparator.h"
+#include "rocksdb/convenience.h"
 #include "rocksdb/env.h"
 #include "rocksdb/filter_policy.h"
 #include "rocksdb/memtablerep.h"
@@ -280,8 +281,6 @@ void ColumnFamilyOptions::Dump(Logger* log) const {
                      rate_limit_delay_max_milliseconds);
     ROCKS_LOG_HEADER(log, "               Options.disable_auto_compactions: %d",
                      disable_auto_compactions);
-    ROCKS_LOG_HEADER(log, "               Options.use_secondary_cache: %d",
-                     use_secondary_cache);
 
     const auto& it_compaction_style =
         compaction_style_to_string.find(compaction_style);
@@ -461,8 +460,8 @@ Options::PrepareForBulkLoad()
   // The compaction would create large files in L1.
   target_file_size_base = 256 * 1024 * 1024;
 
-  // Disable secondary cache by detault
-  use_secondary_cache = true;
+  // Use secondary cache by detault
+  lowest_used_cache_tier = CacheTier::kNonVolatileTier;
   return this;
 }